backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
sqlalchemy==2.0.23
pymysql==1.1.0
redis==5.0.1
httpx==0.25.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
chromadb==0.4.15
sentence-transformers==2.2.2
numpy==1.26.4
python-multipart==0.0.6
websockets==12.0

# --- 以下为架构加固新增的核心依赖 ---
aiofiles==23.2.1
defusedxml==0.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4